requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
openai>=1.60.0
PyPDF2==3.0.1
beautifulsoup4==4.12.3
requests==2.32.3
python-docx==1.1.2
databricks-sql-connector==3.4.0
pydantic==2.9.2
jinja2==3.1.4
aiofiles==24.1.0
python-dotenv==1.0.1
mistune==3.0.2
itsdangerous==2.2.0
gunicorn
